moves system_swap to pytest
