--- system_swap/test/test_system_swap.py
@@ -1,36 +1,0 @@
-# (C) Datadog, Inc. 2010-2017
-# All rights reserved
-# Licensed under Simplified BSD License (see LICENSE)
-
-# 3p
-import mock
-
-from tests.checks.common import AgentCheckTest
-
-class _PSUtilSwapStatsMock(object):
-    def __init__(self, sin, sout):
-        self.sin = sin
-        self.sout = sout
-
-
-ORIG_SWAP_IN = 115332743168
-ORIG_SWAP_OUT = 22920884224
-
-SWAP_IN_INCR = 2
-SWAP_OUT_INCR = 4
-
-MOCK_PSUTIL_SWAP_STATS = [
-    _PSUtilSwapStatsMock(ORIG_SWAP_IN, ORIG_SWAP_OUT),
-    _PSUtilSwapStatsMock(ORIG_SWAP_IN + SWAP_IN_INCR, ORIG_SWAP_OUT + SWAP_OUT_INCR),
-]
-
-class SystemSwapTestCase(AgentCheckTest):
-
-    CHECK_NAME = 'system_swap'
-
-    @mock.patch('psutil.swap_memory', side_effect=MOCK_PSUTIL_SWAP_STATS)
-    def test_system_swap(self, mock_swap_stats):
-        self.run_check_twice({"instances": [{"tags":["optional:tags"]}]}) # Run check twice, sleeping for 1 sec in between\
-        tags = ["optional:tags"]
-        self.assertMetric('system.swap.swapped_in', value=SWAP_IN_INCR, count=1, tags=tags)
-        self.assertMetric('system.swap.swapped_out', value=SWAP_OUT_INCR, count=1, tags=tags)
