+++ system_swap/tox.ini
@@ -1,0 +1,30 @@
+[tox]
+minversion = 2.0
+basepython = py27
+envlist =
+  unit
+  flake8
+
+[testenv]
+platform = linux|darwin|win32
+passenv =
+    DOCKER*
+    COMPOSE*
+deps =
+    -e../datadog_checks_base
+    -e../datadog_checks_base[deps]
+    -rrequirements-dev.txt
+
+[testenv:unit]
+commands =
+    pip install --require-hashes -r requirements.txt
+    pytest -v
+
+[testenv:flake8]
+skip_install = true
+deps = flake8
+commands = flake8 .
+
+[flake8]
+exclude = .eggs,.tox
+max-line-length = 120
